# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AURUM Luxury Fashion eCommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  slug: string;
  name: string;
  brand: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  /** @nullable */
  badge?: string | null;
  colors: string[];
  sizes: string[];
  description?: string;
  figType: string;
  bgGradient: string;
  figColorA: string;
  figColorB: string;
  sku: string;
  stock: number;
  tags: string[];
}